feat(api): add admin authorization
